requirements.txt
streamlit
torch
pillow
transformers